explorer_backend/requirements.txt
django<2.2.0
requests==2.22.0
web3==4.9.2
flask==1.1.0
flask-restful==0.3.7
django_cors_headers==3.0.2
djcacheutils==3.0.0
